+++ packages/multistream-select/src/handle.ts
@@ -2,9 +2,9 @@ import { lpStream } from '@libp2p/utils'
 import { encode } from 'it-length-prefixed'
 import { Uint8ArrayList } from 'uint8arraylist'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
-import { MAX_PROTOCOL_LENGTH, PROTOCOL_ID } from './constants.js'
-import { readString } from './multistream.js'
-import type { MultistreamSelectInit } from './index.js'
+import { MAX_PROTOCOL_LENGTH, PROTOCOL_ID } from './constants.ts'
+import { readString } from './multistream.ts'
+import type { MultistreamSelectInit } from './index.ts'
 import type { MultiaddrConnection, MessageStream } from '@libp2p/interface'
 
 /**