+++ packages/stream-multiplexer-mplex/test/coder.spec.ts
@@ -7,8 +7,8 @@ import { Uint8ArrayList } from 'uint8arraylist'
 import { concat as uint8ArrayConcat } from 'uint8arrays/concat'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { encode } from '../src/encode.js'
-import { decode } from './fixtures/decode.js'
-import { messageWithBytes } from './fixtures/utils.js'
+import { decode } from './fixtures/decode.ts'
+import { messageWithBytes } from './fixtures/utils.ts'
 import type { Message, NewStreamMessage } from '../src/message-types.js'
 
 describe('coder', () => {