+++ packages/libp2p/CHANGELOG.md
@@ -110,6 +110,53 @@
     * @libp2p/tcp bumped from ^9.0.14 to ^9.0.15
     * @libp2p/websockets bumped from ^8.0.14 to ^8.0.15
 
+## [3.2.1](https://github.com/libp2p/js-libp2p/compare/libp2p-v3.2.0...libp2p-v3.2.1) (2026-04-14)
+
+
+### Bug Fixes
+
+* **connection-manager:** add per-address timeout ([#3412](https://github.com/libp2p/js-libp2p/issues/3412)) ([2151144](https://github.com/libp2p/js-libp2p/commit/2151144c3eb6c9a0f3c39123822f8cc22ce3211b))
+* support require ([#3449](https://github.com/libp2p/js-libp2p/issues/3449)) ([f21efd1](https://github.com/libp2p/js-libp2p/commit/f21efd121a9d3755f28049541c1b0758e3498683))
+
+
+### Dependencies
+
+* update p-retry to 8.x.x ([#3442](https://github.com/libp2p/js-libp2p/issues/3442)) ([aaf3fbb](https://github.com/libp2p/js-libp2p/commit/aaf3fbbe224bb7c62dfb9a2ffc8751794a93daec))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.15 to ^5.1.16
+    * @libp2p/interface bumped from ^3.2.0 to ^3.2.1
+    * @libp2p/interface-internal bumped from ^3.1.0 to ^3.1.1
+    * @libp2p/logger bumped from ^6.2.4 to ^6.2.5
+    * @libp2p/multistream-select bumped from ^7.0.15 to ^7.0.16
+    * @libp2p/peer-collections bumped from ^7.0.15 to ^7.0.16
+    * @libp2p/peer-id bumped from ^6.0.6 to ^6.0.7
+    * @libp2p/peer-store bumped from ^12.0.15 to ^12.0.16
+    * @libp2p/utils bumped from ^7.0.15 to ^7.0.16
+
+## [3.2.0](https://github.com/libp2p/js-libp2p/compare/libp2p-v3.1.7...libp2p-v3.2.0) (2026-04-08)
+
+
+### Features
+
+* add DialTarget type ([#3428](https://github.com/libp2p/js-libp2p/issues/3428)) ([2a3e547](https://github.com/libp2p/js-libp2p/commit/2a3e547380a62dffae31f7b5423dfcf1fc0c7433))
+* add high-level connection and stream progress events ([#3429](https://github.com/libp2p/js-libp2p/issues/3429)) ([709b3fa](https://github.com/libp2p/js-libp2p/commit/709b3fa2e94d99c77d97df2dda39b431f8b92878))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.14 to ^5.1.15
+    * @libp2p/interface bumped from ^3.1.1 to ^3.2.0
+    * @libp2p/interface-internal bumped from ^3.0.14 to ^3.1.0
+    * @libp2p/logger bumped from ^6.2.3 to ^6.2.4
+    * @libp2p/multistream-select bumped from ^7.0.14 to ^7.0.15
+    * @libp2p/peer-collections bumped from ^7.0.14 to ^7.0.15
+    * @libp2p/peer-id bumped from ^6.0.5 to ^6.0.6
+    * @libp2p/peer-store bumped from ^12.0.14 to ^12.0.15
+    * @libp2p/utils bumped from ^7.0.14 to ^7.0.15
+
 ## [3.1.7](https://github.com/libp2p/js-libp2p/compare/libp2p-v3.1.6...libp2p-v3.1.7) (2026-03-27)