+++ packages/peer-id/CHANGELOG.md
@@ -11,6 +11,31 @@
   * dependencies
     * @libp2p/interface bumped from ^1.0.1 to ^1.0.2
 
+## [6.0.7](https://github.com/libp2p/js-libp2p/compare/peer-id-v6.0.6...peer-id-v6.0.7) (2026-04-14)
+
+
+### Bug Fixes
+
+* support require ([#3449](https://github.com/libp2p/js-libp2p/issues/3449)) ([f21efd1](https://github.com/libp2p/js-libp2p/commit/f21efd121a9d3755f28049541c1b0758e3498683))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.15 to ^5.1.16
+    * @libp2p/interface bumped from ^3.2.0 to ^3.2.1
+
+## [6.0.6](https://github.com/libp2p/js-libp2p/compare/peer-id-v6.0.5...peer-id-v6.0.6) (2026-04-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.14 to ^5.1.15
+    * @libp2p/interface bumped from ^3.1.1 to ^3.2.0
+
 ## [6.0.5](https://github.com/libp2p/js-libp2p/compare/peer-id-v6.0.4...peer-id-v6.0.5) (2026-03-27)