+++ packages/transport-webtransport/CHANGELOG.md
@@ -96,6 +96,45 @@
   * devDependencies
     * libp2p bumped from ^1.2.2 to ^1.2.3
 
+## [6.0.18](https://github.com/libp2p/js-libp2p/compare/webtransport-v6.0.17...webtransport-v6.0.18) (2026-04-14)
+
+
+### Bug Fixes
+
+* support require ([#3449](https://github.com/libp2p/js-libp2p/issues/3449)) ([f21efd1](https://github.com/libp2p/js-libp2p/commit/f21efd121a9d3755f28049541c1b0758e3498683))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.2.0 to ^3.2.1
+    * @libp2p/peer-id bumped from ^6.0.6 to ^6.0.7
+    * @libp2p/utils bumped from ^7.0.15 to ^7.0.16
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.15 to ^5.1.16
+    * @libp2p/daemon-client bumped from ^10.0.17 to ^10.0.18
+    * @libp2p/logger bumped from ^6.2.4 to ^6.2.5
+    * @libp2p/ping bumped from ^3.1.0 to ^3.1.1
+    * libp2p bumped from ^3.2.0 to ^3.2.1
+
+## [6.0.17](https://github.com/libp2p/js-libp2p/compare/webtransport-v6.0.16...webtransport-v6.0.17) (2026-04-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^3.1.1 to ^3.2.0
+    * @libp2p/peer-id bumped from ^6.0.5 to ^6.0.6
+    * @libp2p/utils bumped from ^7.0.14 to ^7.0.15
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.14 to ^5.1.15
+    * @libp2p/daemon-client bumped from ^10.0.16 to ^10.0.17
+    * @libp2p/logger bumped from ^6.2.3 to ^6.2.4
+    * @libp2p/ping bumped from ^3.0.14 to ^3.1.0
+    * libp2p bumped from ^3.1.7 to ^3.2.0
+
 ## [6.0.16](https://github.com/libp2p/js-libp2p/compare/webtransport-v6.0.15...webtransport-v6.0.16) (2026-03-27)